Move 'SkinConversion' to its own package
