+++ src/com/reco1l/ibancho/data/RoomMods.kt
@@ -1,6 +1,6 @@
 package com.reco1l.ibancho.data
 
-import com.reco1l.osu.conversion.modsToReadable
+import com.reco1l.osu.multiplayer.modsToReadable
 import com.reco1l.osu.multiplayer.Multiplayer
 import ru.nsu.ccfit.zuev.osu.game.mods.GameMod
 import ru.nsu.ccfit.zuev.osu.game.mods.GameMod.*
